IOS-6900 Separate by groupes
