+++ Tangem/Modules/StakingDetails/StakingDetailsView.swift
@@ -37,7 +37,7 @@ struct StakingDetailsView: View {
                 actionButton
             }
             .background(Colors.Background.secondary)
-            .navigationTitle("Staking Solana")
+            .navigationTitle(viewModel.title)
             .navigationBarTitleDisplayMode(.inline)
         }
     }
@@ -54,7 +54,7 @@ struct StakingDetailsView: View {
         GroupedSection(viewModel.averageRewardingViewData) {
             AverageRewardingView(data: $0)
         } header: {
-            DefaultHeaderView("Average Reward Rate")
+            DefaultHeaderView(Localization.stakingDetailsAverageRewardRate)
         }
         .interItemSpacing(12)
         .innerContentPadding(12)
@@ -64,104 +64,19 @@ struct StakingDetailsView: View {
         GroupedSection(viewModel.rewardViewData) {
             RewardView(data: $0)
         } header: {
-            DefaultHeaderView("Rewards")
+            DefaultHeaderView(Localization.stakingRewards)
         }
         .interItemSpacing(12)
         .innerContentPadding(12)
     }
 
     private var actionButton: some View {
-        MainButton(title: "Stake") {
+        MainButton(title: Localization.commonStake) {
             viewModel.userDidTapActionButton()
         }
         .padding(.horizontal, 16)
         .padding(.bottom, 8)
         .readGeometry(\.size.height, bindTo: $bottomViewHeight)
-    }
-}
-
-struct RewardViewData: Hashable, Identifiable {
-    var id: Int { hashValue }
-
-    let state: State
-}
-
-extension RewardViewData {
-    enum State: Hashable {
-        case noRewards
-        case rewards(fiatFormatted: String, cryptoFormatted: String)
-    }
-}
-
-struct RewardView: View {
-    let data: RewardViewData
-
-    var body: some View {
-        HStack(spacing: 4) {
-            content
-        }
-        .infinityFrame(axis: .horizontal, alignment: .leading)
-    }
-
-    @ViewBuilder
-    var content: some View {
-        switch data.state {
-        case .noRewards:
-            Text("No rewards to claim")
-                .style(Fonts.Regular.subheadline, color: Colors.Text.tertiary)
-        case .rewards(let fiatFormatted, let cryptoFormatted):
-            Text(fiatFormatted)
-                .style(Fonts.Regular.subheadline, color: Colors.Text.primary1)
-
-            Text(AppConstants.dotSign)
-                .style(Fonts.Regular.subheadline, color: Colors.Text.primary1)
-
-            Text(cryptoFormatted)
-                .style(Fonts.Regular.subheadline, color: Colors.Text.tertiary)
-        }
-    }
-}
-
-struct AverageRewardingViewData: Hashable, Identifiable {
-    var id: Int { hashValue }
-
-    let aprFormatted: String
-    let profitFormatted: String
-}
-
-struct AverageRewardingView: View {
-    let data: AverageRewardingViewData
-
-    var body: some View {
-        HStack(spacing: .zero) {
-            leadingView
-
-            trailingView
-        }
-        .infinityFrame(axis: .horizontal, alignment: .leading)
-    }
-
-    private var leadingView: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            Text("APR")
-                .style(Fonts.Regular.footnote, color: Colors.Text.tertiary)
-
-            Text(data.aprFormatted)
-                .style(Fonts.Regular.callout, color: Colors.Text.accent)
-        }
-        .infinityFrame(axis: .horizontal, alignment: .leading)
-    }
-
-    private var trailingView: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            Text("30 day est. profit")
-                .style(Fonts.Regular.footnote, color: Colors.Text.tertiary)
-
-            Text(data.profitFormatted)
-                .style(Fonts.Regular.callout, color: Colors.Text.primary1)
-        }
-        .lineLimit(1)
-        .infinityFrame(axis: .horizontal, alignment: .leading)
     }
 }
 
@@ -185,6 +100,7 @@ struct StakingDetailsView_Preview: PreviewProvider {
             unbonding: .days(3),
             minimumRequirement: 0.000028,
             rewardClaimingType: .auto,
+            rewardType: .apr,
             warmupPeriod: .days(3),
             rewardScheduleType: .block
         ),
